% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pasta}
\alias{pasta}
\title{pasta}
\usage{
pasta(arglist = arglist_get(...))
}
\arguments{
\item{...}{Arguments}
}
\description{
Run pasta
}
\examples{
library(outsider)
pasta <- module_import('pasta', repo = 'dombennett/om..pasta')
pasta('-h')
}
